requirements.txt
asgiref==3.7.2
boto3==1.34.49
botocore==1.34.49
channels==4.0.0
click==8.1.7
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
django-storages==1.14.2
gunicorn==21.2.0
h11==0.14.0
jmespath==1.0.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
s3transfer==0.10.0
send-email==201212
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
uvicorn==0.25.0
whitenoise==6.6.0
